Move add_eval_data to BaseDocumentStore
